backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
python-dotenv==1.0.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
alembic==1.13.1

# PDF Processing
PyPDF2==3.0.1
pdf2image==1.17.0
pytesseract==0.3.10
reportlab==4.0.9
Pillow==10.2.0

# Document Generation
python-docx==1.1.0
openpyxl==3.1.2

# AI/ML
google-generativeai==0.3.2
langchain==0.1.0
langchain-google-genai==0.0.6

# Utilities
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1

# Logging and Monitoring
loguru==0.7.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
